#116: refactor code
