+++ restservice/src/main/java/org/itevents/controller/UserRestController.java
@@ -8,9 +8,8 @@ import org.itevents.model.Event;
 import org.itevents.model.Filter;
 import org.itevents.model.User;
 import org.itevents.model.builder.UserBuilder;
-import org.itevents.security.AESCryptTokenService;
-import org.itevents.security.CryptTokenService;
-import org.itevents.security.Token;
+import org.itevents.service.CryptTokenService;
+import org.itevents.service.security.Token;
 import org.itevents.service.EventService;
 import org.itevents.service.FilterService;
 import org.itevents.service.RoleService;
