+++ restservice/src/main/java/org/itevents/service/security/RestAuthenticationEntryPoint.java
@@ -1,17 +1,12 @@
-package org.itevents.security;
+package org.itevents.service.security;
 
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Collection;
 
 /**
  * The Entry Point will not redirect to any sort of Login - it will return the 401
