+++ restservice/src/main/resources/spring-security.xml
@@ -10,7 +10,7 @@
 		http://www.springframework.org/schema/context/spring-context.xsd">
 
     <context:property-placeholder location="classpath*:local.properties"/>
-    <context:component-scan base-package="org.itevents.security"/>
+    <context:component-scan base-package="org.itevents.service.security"/>
     <global-method-security pre-post-annotations="enabled"/>
     <http realm="itevents"
           entry-point-ref="restAuthenticationEntryPoint"
@@ -30,12 +30,12 @@
         <intercept-url pattern="/users/me" access='isAuthenticated()'/>
         <intercept-url pattern="/**" access="permitAll()"/>
 
-        <custom-filter ref="basicAuthenticationFilter" position="FORM_LOGIN_FILTER"   />
+        <custom-filter ref="jwtAuthenticationFilter" position="FORM_LOGIN_FILTER"   />
         <anonymous username="guest"/>
     </http>
 
-    <beans:bean id="basicAuthenticationFilter"
-          class="org.itevents.security.CustomTokenAuthenticationFilter">
+    <beans:bean id="jwtAuthenticationFilter"
+          class="org.itevents.service.security.CustomTokenAuthenticationFilter">
         <beans:constructor-arg index="0" ref="authenticationManager"/>
     </beans:bean>
